requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
openpyxl
onnxruntime
sqlalchemy
fpdf2
requests
joblib
